+++ src/pages/species-richness/components/species-richness-table/species-richness-table.ts
@@ -1,8 +1,14 @@
-import { Vue } from 'vue-class-component'
+import { Options, Vue } from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
 
 import { ChartModels } from '@/models'
+import ExportButtonView from '@/views/export-button.vue'
 
+@Options({
+  components: {
+    ExportButtonView
+  }
+})
 export default class SpeciesRichnessTable extends Vue {
   @Prop({ default: [] }) tableData!: ChartModels.TableData[]
 
@@ -10,11 +16,11 @@ export default class SpeciesRichnessTable extends Vue {
     return [
       'Species',
       'Class',
-      ...(this.hasData ? Object.keys(this.tableData[0]).filter(k => !['speciesName', 'className'].includes(k)) : [])
+      ...(this.hasTableData ? Object.keys(this.tableData[0]).filter(k => !['speciesName', 'className'].includes(k)) : [])
     ]
   }
 
-  get hasData (): boolean {
+  get hasTableData (): boolean {
     return this.tableData.length > 0
   }
 }