+++ CHANGELOG.md
@@ -19,4 +19,5 @@
 * **features:** Map to be export as png (#86)
 * **features:** Map style toggle (#88)
 * **features:** Bar chart support responsive (#61)
-* **features:** Table - list species name in the table(#82)
+* **features:** Table - list species name in the table (#82)
+* **features:** Export species richness raw data as csv (#36)